Leetcode Problem 3: Lenght of Longest Substring
Description:
https://leetcode.com/problems/longest-substring-without-repeating-characters/

Solution stats:
Runtime = 28ms -> 32.02 percentile -> O(N^2)
Memory = 5.4MB -> 100 percentile

Explanantion:
We traverse the fullString using two pointers - head and tail.
We maintain a character array of size 128 to store substrings without
repeating characters. 128 characters size is sufficeint because there is
bound to be a repeating character by the time we reach that size

Head moves forward through the fullString. If character pointed to by
Head is not present inside subString, add it

If character pointed to by Head is present in substring, move Tail till
the position after first occurence of this character. Note: Use
substring to calculate the position to move tai.
Flush the entire substring and update substring using currnt head and
tail positions in the fullString